TRIFFHIR-372 adding the dropdown and typeahead for selecting publisher version to export screen
