+++ apps/client/src/app/export/export.component.html
@@ -52,15 +52,7 @@
         </div>
 
         <div class="col-md-6" *ngIf="options.includeIgPublisherJar">
-          <app-fhir-boolean
-            [parentObject]="options"
-            propertyName="useLatest"
-            noLabel="No ({{Globals.packagedIgPublisherVersion}})"
-            [required]="true"
-            [cookieKey]="Globals.cookieKeys.exportLastUseLatestIgPublisher"
-            title="Use latest FHIR IG Publisher?">
-          </app-fhir-boolean>
-          <div class="help-block">The latest version of the FHIR IG Publisher may include changes that are not supported or not compatible with the export from Trifolia-on-FHIR. If Publish fails to run, try re-running with 'No' selected for this option.</div>
+          <trifolia-fhir-ig-publisher-selection [options]="options"></trifolia-fhir-ig-publisher-selection>
         </div>
       </div>
       <div class="row">
