+++ apps/client/src/app/publish/publish.component.ts
@@ -69,41 +69,6 @@ export class PublishComponent implements OnInit {
     });
   }
 
-  /**
-   * this metehod is called when the user types into the typeahead input box for IG Publisher version
-   * @param $e
-   */
-  typeaheadSelectedVersion($e) {
-    $e.preventDefault();
-    const item = $e.item;
-    this.versionTypeahead = item;
-    this.versionDropdown = item;
-    const version = item.replace(' (Current)', '');
-    this.options.version = version;
-  }
-
-  /**
-   * this method is called when the user selects an item in the IG Publisher top 10 versions drop down
-   * @param $event
-   */
-  dropdownSelectedVersion($event){
-    $event.preventDefault();
-    const item = $event.target.options[$event.target.options.selectedIndex].text;
-    console.log(item);
-    this.versionDropdown = item;
-    this.versionTypeahead = item;
-    const version = item.replace(' (Current)', '');
-    this.options.version = version;
-  }
-
-  search = (text$: Observable<string>) =>
-    text$.pipe(
-      debounceTime(200),
-      distinctUntilChanged(),
-      map(term => term.length < 2 ? []
-        : this.publisherVersions.filter(v => v.toLowerCase().indexOf(term.toLowerCase()) > -1).slice(0, 10))
-    )
-
   public implementationGuideChanged(implementationGuide: ImplementationGuide) {
     this.selectedImplementationGuide = implementationGuide;
     this.options.implementationGuideId = implementationGuide ? implementationGuide.id : undefined;
@@ -126,18 +91,6 @@ export class PublishComponent implements OnInit {
           (err) => this.message = getErrorString(err)
         );
     }
-
-    this.http.get('/api/export/publisher-version')
-      .subscribe(data => {
-          this.publisherVersions = data;
-          this.recentPublisherVersions = this.publisherVersions.splice(0,10);
-          this.versionTypeahead = this.recentPublisherVersions[0];
-          const version = this.recentPublisherVersions[0].replace(' (Current)', '');
-          this.options.version = version;
-        },
-        error => {
-          console.error(error);
-        });
   }
 
   public searchImplementationGuide = (text$: Observable<string>) => {
