+++ apps/client/src/app/shared-ui/ig-publisher-selection/ig-publisher-selection.component.html
@@ -1,0 +1,13 @@
+
+<label for="version-typeahead">FHIR IG Publisher Version.</label>
+<div class="input-group">
+  <input id="version-typeahead" type="text" class="form-control" (selectItem)="typeaheadSelectedVersion($event)" [(ngModel)]="versionTypeahead" [ngbTypeahead]="search"/>
+  <div class="input-group-append">
+    <select id='versions' name='versions' class="form-control" (change)="dropdownSelectedVersion($event)" [(ngModel)]="versionDropdown">
+      <option *ngFor="let version of recentPublisherVersions">
+        {{version}}
+      </option>
+    </select>
+  </div>
+  <div class="help-block">The latest version of the FHIR IG Publisher may include changes that are not supported or not compatible with the export from Trifolia-on-FHIR. If Publish fails to run, try re-running with 'No' selected for this option.</div>
+</div>
