+++ apps/server/src/app/export.controller.ts
@@ -174,7 +174,7 @@ export class ExportController extends BaseController {
       implementationGuideId);
 
     try {
-      await exporter.export(options.format, options.includeIgPublisherJar, options.includeIgPublisherJar ? options.useLatest : false, options.templateType, options.template, options.templateVersion);
+      await exporter.export(options.format, options.includeIgPublisherJar, options.includeIgPublisherJar ? options.version : false, options.templateType, options.template, options.templateVersion);
 
       // Zip up the dir, send it to client, and then delete the directory
       await this.sendPackageResponse(exporter.packageId, response);
